Refactor sendAgentSensations
Separated sendAgentSensations so that the loop over the GOBAgent collection
takes place and then calls a method sendSensationsToAgent for any agents
that need a sensory update. This was intended to separate the potential conflicts
calling sendAgentSensetations within the AgentListener, which now calls
sendSensationsToAgent directly for the new agent that was just created.