Fix Issue w/ Quicksand
If agents did not send an action on the current simulator cycle, we were actually performing
a "wait" action, which if on quicksand will naturally kill the agent. Added a method
GOBAgent.decAgentEnergy(decrement) that is called in Grid.processAgentActions instead
of enacting a wait action. Also changed all decrements of agentEnergy inside GOBAgent so
they use the decAgentEnergy method.